Move display package to nextjs fork
