+++ nextjs/packages/next/package.json
@@ -222,6 +222,7 @@
     "comment-json": "3.0.3",
     "compression": "1.7.4",
     "conf": "5.0.0",
+    "console-table-printer": "2.7.5",
     "content-type": "1.0.4",
     "cookie": "0.4.1",
     "css-loader": "4.3.0",
