+++ nextjs/packages/next/stdlib-server/auth-sessions.ts
@@ -1,5 +1,4 @@
-// Must import this type from 'blitz'
-// import { log } from '@blitzjs/display'
+import { log } from '../server/lib/logging'
 import { fromBase64, toBase64 } from 'b64-lite'
 import cookie from 'next/dist/compiled/cookie'
 import fs from 'fs'
@@ -95,8 +94,7 @@ const defaultConfig: SessionConfig = {
     } catch (error) {
       // Session doesn't exist in DB for some reason, so create it
       if (error.code === 'P2016') {
-        // log.warning("Could not update session because it's not in the DB")
-        console.log("Could not update session because it's not in the DB")
+        log.warning("Could not update session because it's not in the DB")
       } else {
         throw error
       }
@@ -700,8 +698,7 @@ async function getSessionKernel(
       persistedSession.antiCSRFToken !== antiCSRFToken
     ) {
       if (!antiCSRFToken) {
-        // log.warning(
-        console.log(
+        log.warning(
           `This request is missing the ${HEADER_CSRF} header. You can learn about adding this here: https://blitzjs.com/docs/session-management#manual-api-requests`
         )
       }
@@ -777,8 +774,7 @@ async function getSessionKernel(
 
     if (enableCsrfProtection && payload.antiCSRFToken !== antiCSRFToken) {
       if (!antiCSRFToken) {
-        // log.warning(
-        console.log(
+        log.warning(
           `This request is missing the ${HEADER_CSRF} header. You can learn about adding this here: https://blitzjs.com/docs/session-management#manual-api-requests`
         )
       }
