+++ packages/blitz/package.json
@@ -53,7 +53,6 @@
   "dependencies": {
     "@blitzjs/babel-preset": "0.42.4",
     "@blitzjs/cli": "0.42.4",
-    "@blitzjs/display": "0.42.4",
     "@blitzjs/generator": "0.42.4",
     "@blitzjs/server": "0.42.4",
     "@testing-library/jest-dom": "5.11.9",
