+++ packages/cli/src/commands/console.ts
@@ -19,7 +19,7 @@ export class Console extends Command {
   async run() {
     process.env.CLI_COMMAND_CONSOLE = "true"
     this.parse(Console)
-    const {log} = require("@blitzjs/display")
+    const {log} = require("next/dist/server/lib/logging")
     const chalk = require("chalk")
     log.branded("You have entered the Blitz console")
     console.log(chalk.yellow("Tips: - Exit by typing .exit or pressing Ctrl-D"))
