+++ packages/cli/src/commands/db.ts
@@ -1,6 +1,6 @@
-import {log} from "@blitzjs/display"
 import {Command, flags} from "@oclif/command"
 import {baseLogger} from "next/dist/server/lib/logging"
+import {log} from "next/dist/server/lib/logging"
 
 export function getDbName(connectionString: string): string {
   const dbUrlParts: string[] = connectionString!.split("/")
