+++ packages/cli/src/commands/new.ts
@@ -1,10 +1,10 @@
-import {log} from "@blitzjs/display"
 import type {AppGeneratorOptions} from "@blitzjs/generator"
 import {getLatestVersion} from "@blitzjs/generator"
 import {flags} from "@oclif/command"
 import chalk from "chalk"
 import spawn from "cross-spawn"
 import hasbin from "hasbin"
+import {log} from "next/dist/server/lib/logging"
 import {lt} from "semver"
 const debug = require("debug")("blitz:new")
