--- packages/display/.gitignore
@@ -1,4 +1,0 @@
-*.log
-.DS_Store
-node_modules
-dist
