--- packages/display/README.md
@@ -1,3 +1,0 @@
-# `display`
-
-This package houses all files related to Blitz commandline output.
