--- packages/display/jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  preset: "../../jest-unit.config.js",
-}
