--- packages/display/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@blitzjs/display",
-  "version": "0.42.4",
-  "description": "Display package for the Blitz CLI",
-  "homepage": "https://github.com/blitz-js/blitz#readme",
-  "license": "MIT",
-  "scripts": {
-    "test": "jest",
-    "test:watch": "jest --watch"
-  },
-  "author": {
-    "name": "Brandon Bayer",
-    "email": "b@bayer.ws",
-    "url": "https://twitter.com/flybayer"
-  },
-  "main": "dist/blitzjs-display.cjs.js",
-  "module": "dist/blitzjs-display.esm.js",
-  "types": "dist/blitzjs-display.cjs.d.ts",
-  "files": [
-    "dist"
-  ],
-  "keywords": [
-    "blitz",
-    "display"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/blitz-js/blitz.git"
-  },
-  "dependencies": {
-    "@blitzjs/display": "0.42.4",
-    "chalk": "^4.1.0",
-    "console-table-printer": "^2.7.5",
-    "ora": "^5.3.0",
-    "readline": "1.3.0"
-  }
-}
