--- packages/display/src/index.test.ts
@@ -1,5 +1,0 @@
-export {}
-
-it("should do nothing", () => {
-  expect(true).toBe(true)
-})
