+++ packages/generator/src/generators/model-generator.ts
@@ -1,7 +1,7 @@
-import {log} from "@blitzjs/display"
 import * as ast from "@mrleebo/prisma-ast"
 import {spawn} from "cross-spawn"
 import {newline} from "next/dist/server/lib/logging"
+import {log} from "next/dist/server/lib/logging"
 import which from "npm-which"
 import path from "path"
 import {Generator, GeneratorOptions, SourceRootType} from "../generator"
