+++ packages/server/package.json
@@ -19,7 +19,6 @@
     "dist"
   ],
   "dependencies": {
-    "@blitzjs/display": "0.42.4",
     "cross-spawn": "7.0.3",
     "detect-port": "1.3.0",
     "esbuild": "^0.11.12",
