+++ packages/server/src/next-utils.ts
@@ -1,9 +1,9 @@
-import {log} from "@blitzjs/display"
 import {ChildProcess} from "child_process"
 import {spawn} from "cross-spawn"
 import detect from "detect-port"
 import * as esbuild from "esbuild"
 import {existsSync, readJSONSync} from "fs-extra"
+import {log} from "next/dist/server/lib/logging"
 import {newline} from "next/dist/server/lib/logging"
 import {getProjectRootSync} from "next/dist/server/lib/utils"
 import path from "path"
